requirements.txt
Django==3.1.2
django-crispy-forms==1.9.2
python-decouple==3.3
SQLAlchemy==1.3.20
